% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_12.R
\name{as_C12}
\alias{as_C12}
\title{Change a vector into object of C12 group}
\usage{
as_C12(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to C12
}
\description{
Change a vector into object of C12 group
}
\examples{
x <- 0:7
as_C12(x)
}
