% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3C2 group.R, R/C3C2.R
\name{as_C3C2}
\alias{as_C3C2}
\title{Change a vector into object of C3C2 group}
\usage{
as_C3C2(x)

as_C3C2(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to C3C2

An Sgroupr_sgrp group belonging to C3C2
}
\description{
Change a vector into object of C3C2 group

Change a vector into object of C3C2 group
}
\examples{
x <- 0:5
as_C3C2(x)
x <- 0:5
as_C3C2(x)
}
