% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C4.R
\name{as_C4_C4}
\alias{as_C4_C4}
\title{Change a vector into object of C4_C4 group}
\usage{
as_C4_C4(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to C4_C4
}
\description{
Change a vector into object of C4_C4 group
}
\examples{
x <- 0:15
as_C4_C4(x)
}
