% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8_C2.R
\name{as_C8_C2}
\alias{as_C8_C2}
\title{Change a vector into object of C8_C2 group}
\usage{
as_C8_C2(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to C8_C2
}
\description{
Change a vector into object of C8_C2 group
}
\examples{
x <- 0:15
as_C8_C2(x)
}
