% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_16.R
\name{as_Cyclic_16}
\alias{as_Cyclic_16}
\title{Change a vector into object of Cyclic 16 group}
\usage{
as_Cyclic_16(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Cyclic 16
}
\description{
Change a vector into object of Cyclic 16 group
}
\examples{
x <- 0:15
as_Cyclic_16(x)
}
