% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_3.R
\name{as_Cyclic_3}
\alias{as_Cyclic_3}
\title{Change a vector into object of Cyclic_3 group}
\usage{
as_Cyclic_3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Cyclic_3
}
\description{
Change a vector into object of Cyclic_3 group
}
\examples{
x <- 0:7
as_Cyclic_3(x)
}
