% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic4.R
\name{as_Cyclic_4}
\alias{as_Cyclic_4}
\title{Change a vector into object of Cyclic_4 group}
\usage{
as_Cyclic_4(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Cyclic_4
}
\description{
Change a vector into object of Cyclic_4 group
}
\examples{
x <- 0:3
as_Cyclic_4(x)
}
