% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_8.R
\name{as_Cyclic_8}
\alias{as_Cyclic_8}
\title{Change a vector into object of Cyclic 8 group}
\usage{
as_Cyclic_8(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Cyclic 8
}
\description{
Change a vector into object of Cyclic 8 group
}
\examples{
x <- 0:7
as_Cyclic_8(x)
}
