% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dicyclic8.R
\name{as_Dcyclic_8}
\alias{as_Dcyclic_8}
\title{Change a vector into object of Dcyclic_8 group}
\usage{
as_Dcyclic_8(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Dcyclic_8
}
\description{
Change a vector into object of Dcyclic_8 group
}
\examples{
x <- 0:7
as_Dcyclic_8(x)
}
