% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_6.R
\name{as_Dih_6}
\alias{as_Dih_6}
\title{Change a vector into object of Dih_6 group}
\usage{
as_Dih_6(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Dih_6
}
\description{
Change a vector into object of Dih_6 group
}
\examples{
x <- 0:7
as_Dih_6(x)
}
