% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_8.R
\name{as_Dih_8}
\alias{as_Dih_8}
\title{Change a vector into object of dihedral group Dih8 (Heisenberg) group}
\usage{
as_Dih_8(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to dihedral group Dih8 (Heisenberg)
}
\description{
Change a vector into object of dihedral group Dih8 (Heisenberg) group
}
\examples{
x <- 0:7
as_Dih_8(x)
}
