% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_C4_C3.R
\name{as_Group_C4_C3}
\alias{as_Group_C4_C3}
\title{Change a vector into object of Group_C4_C3}
\usage{
as_Group_C4_C3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Group_C4_C3
}
\description{
Change a vector into object of Group_C4_C3
}
\examples{
x <- 0:11
as_Group_C4_C3(x)
}
