% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K4_C3.R
\name{as_K4_C3}
\alias{as_K4_C3}
\title{Change a vector into object of K4_C3 group}
\usage{
as_K4_C3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to K4_C3
}
\description{
Change a vector into object of K4_C3 group
}
\examples{
x <- 0:7
as_K4_C3(x)
}
