% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_4.R
\name{as_Klein_4}
\alias{as_Klein_4}
\title{Change a vector into object of Klein_4 group}
\usage{
as_Klein_4(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Klein_4
}
\description{
Change a vector into object of Klein_4 group
}
\examples{
x <- 0:3
as_Klein_4(x)
}
