% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_8 group.R
\name{as_Klein_8}
\alias{as_Klein_8}
\title{Change a vector into object of Klein_8 group}
\usage{
as_Klein_8(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Klein_8
}
\description{
Change a vector into object of Klein_8 group
}
\examples{
x <- 0:7
as_Klein_8(x)
}
