% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{as_S3}
\alias{as_S3}
\title{Change a vector into object of S3 group}
\usage{
as_S3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to S3
}
\description{
Change a vector into object of S3 group
}
\examples{
x <- 0:7
as_S3(x)
}
