% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C12.R
\name{C12}
\alias{C12}
\title{Create Cyclic 12 group}
\usage{
C12(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
A Cyclic 12 group of Sgroupr_sgrp object
}
\description{
Create Cyclic 12 group
}
\examples{
x <- C12(0:11)
}
