% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_C2.R
\name{C3_C2}
\alias{C3_C2}
\title{Create a C3C2 sgroupr_sgrp object}
\usage{
C3_C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C6C2
}
\description{
Create a C3C2 sgroupr_sgrp object
}
\examples{
C3_C2(0:5)
}
