% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3C2 group.R, R/C3C2.R
\name{C3C2}
\alias{C3C2}
\title{Create a C3C2 sgroupr_sgrp object}
\usage{
C3C2(x = integer())

C3C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C6C2

a sgroupr_sgpr object with group C6C2
}
\description{
Create a C3C2 sgroupr_sgrp object

Create a C3C2 sgroupr_sgrp object
}
\examples{
C3C2(0:5)
C3C2(0:5)
}
