% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\name{C4}
\alias{C4}
\title{Create a C4 sgroupr_sgrp object}
\usage{
C4(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C4
}
\description{
Create a C4 sgroupr_sgrp object
}
\examples{
C4(0:4)
}
