% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C4.R
\name{C4_C4}
\alias{C4_C4}
\title{Create a C4_C4 sgroupr_sgrp object}
\usage{
C4_C4(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C4_C4
}
\description{
Create a C4_C4 sgroupr_sgrp object
}
\examples{
C4_C4(0:15)
}
