% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4C2 group.R
\name{C4C2}
\alias{C4C2}
\title{Create a C4C2 sgroupr_sgrp object}
\usage{
C4C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C4C2
}
\description{
Create a C4C2 sgroupr_sgrp object
}
\examples{
C4C2(0:7)
}
