% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6_C2.R
\name{C6_C2}
\alias{C6_C2}
\title{Create a C6_C2 sgroupr_sgrp object}
\usage{
C6_C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C6_C2
}
\description{
C6_C2 belongs to the multiplication Group
}
\examples{
C6_C2(0:11)
}
