% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6C2.R
\name{C6C2}
\alias{C6C2}
\title{Create a C6C2 sgroupr_sgrp object}
\usage{
C6C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C6C2
}
\description{
Create a C6C2 sgroupr_sgrp object
}
\examples{
C6C2(0:11)
}
