% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8_C2.R
\name{C8_C2}
\alias{C8_C2}
\title{Create a C8_C2 sgroupr_sgrp object}
\usage{
C8_C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C8_C2
}
\description{
Create a C8_C2 sgroupr_sgrp object
}
\examples{
C8_C2(0:17)
}
