% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_group.R
\name{check_group}
\alias{check_group}
\title{Check if the group has been construct}
\usage{
check_group(group)
}
\arguments{
\item{group}{The group the small group is}
}
\value{
True if the specified group is included inside our package and False
if it has not.
}
\description{
Check if the group has been construct
}
\examples{
check_group("Klein_4")
}
