% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_number.R
\name{check_number}
\alias{check_number}
\title{Check the number of specific group}
\usage{
check_number(x, group)
}
\arguments{
\item{x}{An integer vector}

\item{group}{The group we are going to create}
}
\value{
A vector that change
}
\description{
Check the number of specific group
}
\examples{
check_number(0:7, group = "S3")
}
