% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Cayley table for finite small group}
\usage{
create_table(x)
}
\arguments{
\item{x}{Is an sgroupr_sgrp object}
}
\value{
A matrix for arithmetic operation
}
\description{
Cayley table for finite small group
}
\examples{
x <- Cyclic_8(0:7)
create_table(x)
}
