% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_12.R
\name{Cyclic_12}
\alias{Cyclic_12}
\title{Create Cyclic 12 group}
\usage{
Cyclic_12(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
A Cyclic 12 group of Sgroupr_sgrp object
}
\description{
Create Cyclic 12 group
}
\examples{
x <- Cyclic_12(0:11)
}
