% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_3.R
\name{Cyclic_3}
\alias{Cyclic_3}
\title{Create a Cyclic 3 sgroupr_sgrp object}
\usage{
Cyclic_3(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with group Cyclic_3
}
\description{
The length of Cyclic 3 should be 0 to 2
}
