% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_4 group.R
\name{Cyclic_4}
\alias{Cyclic_4}
\title{Create a Cyclic_4 sgroupr_sgrp object}
\usage{
Cyclic_4(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group Cyclic_4
}
\description{
Create a Cyclic_4 sgroupr_sgrp object
}
\examples{
Cyclic_4(0:3)
}
