% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_6.R
\name{Cyclic_6}
\alias{Cyclic_6}
\title{Create a Cyclic 6 sgroupr_sgrp object}
\usage{
Cyclic_6(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with group Cyclic 6
}
\description{
The length of Cyclic 6 should be 0 to 5
}
\examples{
Cyclic_6(0:5)
}
