% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_8.R
\name{Cyclic_8}
\alias{Cyclic_8}
\title{Create a Cyclic 8 sgroupr_sgrp object}
\usage{
Cyclic_8(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with group Cyclic 8
}
\description{
Cyclic 8 belongs to the Cyclic Group
}
\examples{
Cyclic_8(0:7)
}
