% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dcyclic_8.R
\name{Dcyclic_8}
\alias{Dcyclic_8}
\title{Create a Dcyclic_8 sgroupr_sgrp object}
\usage{
Dcyclic_8(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group Dcyclic_8
}
\description{
Dcyclic 8 belongs to the Dicyclic Group
}
\examples{
Dcyclic_8(0:7)
}
