% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih12.R
\name{Dih12}
\alias{Dih12}
\title{Helper function of Dih2 group}
\usage{
Dih12(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a Dih12 sgroupr_sgrp
}
\description{
Helper function of Dih2 group
}
\examples{
Dih12(0:11)
}
