% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih8_square.R
\name{Dih8}
\alias{Dih8}
\title{Create a Dih8 small group of sgroupr_sgrp}
\usage{
Dih8(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
Return an object of sgroupr_sgrp with group Dih8
}
\description{
Create a Dih8 small group of sgroupr_sgrp
}
\examples{
Dih8(0:7)
}
