% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_12.R
\name{Dih_12}
\alias{Dih_12}
\title{Create a Dih_12 sgroupr_sgrp object}
\usage{
Dih_12(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a Dih_12 sgroupr_sgrp
}
\description{
Dih_12 belongs to the dihedral Group
}
\examples{
Dih_12(0:11)
}
