% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_6.R
\name{Dih_6}
\alias{Dih_6}
\title{Create a Dih_6 small group of sgroupr_sgrp}
\usage{
Dih_6(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
An object of sgroupr_sgrp with group Dih_6
}
\description{
Dih_6 belongs to the dihedral Group
}
\examples{
Dih_6(0:5)
}
