% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_8.R
\name{Dih_8}
\alias{Dih_8}
\title{Create a dihedral group Dih8 (Heisenberg) sgroupr_sgrp object}
\usage{
Dih_8(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with dihedral group Dih_8 (Heisenberg)
}
\description{
Dih_8 (Heisenberg) belongs to the dihedral Group
}
\examples{
Dih_8(0:7)
}
