% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrp.R
\name{group}
\alias{group}
\title{Extract the group attribute of a sgroupr_sgrp}
\usage{
group(x)
}
\arguments{
\item{x}{is an sgroupr_sgrp object}
}
\description{
Extract the group attribute of a sgroupr_sgrp
}
\examples{
x <- Cyclic_3(c(0,1,2))
group(x)
}
