% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c12.R
\name{is_C12}
\alias{is_C12}
\title{Check if a sgroupr_sgpr is C12}
\usage{
is_C12(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C12
}
\examples{
is_C12(C12(0:3))
}
