% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\name{is_C4}
\alias{is_C4}
\title{Check if a sgroupr_sgpr is C4}
\usage{
is_C4(x)
}
\arguments{
\item{An}{sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C4
}
\examples{
is_C4(C4(0:3))
}
