% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C2.R
\name{is_C4_C2}
\alias{is_C4_C2}
\title{Check if a sgroupr_sgpr is C4_C2}
\usage{
is_C4_C2(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C4_C2
}
\examples{
is_C4_C2(C4_C2(0:7))
}
