% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C3.R
\name{is_C4_C3}
\alias{is_C4_C3}
\title{Check if a sgroupr_sgpr is C4_C3}
\usage{
is_C4_C3(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C4_C3
}
\examples{
is_C4_C3(C4_C3(0:11))
}
