% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8C2.R
\name{is_C8C2}
\alias{is_C8C2}
\title{Check if a sgroupr_sgpr is C8C2}
\usage{
is_C8C2(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C8C2
}
\examples{
is_C8C2(C8C2(0:17))
}
