% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_12.R
\name{is_Cyclic_12}
\alias{is_Cyclic_12}
\title{Check if a sgroupr_sgpr is C12}
\usage{
is_Cyclic_12(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C12
}
\examples{
is_Cyclic_12(Cyclic_12(0:11))
}
