% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_16.R
\name{is_Cyclic_16}
\alias{is_Cyclic_16}
\title{Check if a sgroupr_sgpr is Cyclic 16}
\usage{
is_Cyclic_16(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Cyclic 16
}
\examples{
is_Cyclic_16(Cyclic_16(0:15))
}
