% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_3.R
\name{is_Cyclic_3}
\alias{is_Cyclic_3}
\title{Check if a sgroupr_sgpr is Cyclic 3}
\usage{
is_Cyclic_3(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Cyclic 3
}
\examples{
is_Cyclic_3(Cyclic_3(0:2))
}
