% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic4.R
\name{is_Cyclic_4}
\alias{is_Cyclic_4}
\title{Check if a sgroupr_sgpr is Cyclic_4}
\usage{
is_Cyclic_4(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Cyclic_4
}
\examples{
is_Cyclic_4(Cyclic_4(0:3))
}
