% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_6.R
\name{is_Cyclic_6}
\alias{is_Cyclic_6}
\title{Check if a sgroupr_sgpr is Cyclic 6}
\usage{
is_Cyclic_6(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Cyclic 6
}
\examples{
is_Cyclic_6(Cyclic_6(0:5))
}
