% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_8.R
\name{is_Cyclic_8}
\alias{is_Cyclic_8}
\title{Check if a sgroupr_sgpr is Cyclic 8}
\usage{
is_Cyclic_8(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Cyclic 8
}
\examples{
is_Cyclic_8(Cyclic_8(0:7))
}
