% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dcyclic_8.R
\name{is_Dcyclic_8}
\alias{is_Dcyclic_8}
\title{Check if a sgroupr_sgpr is Dcyclic_8}
\usage{
is_Dcyclic_8(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Dcyclic_8
}
\examples{
is_Dcyclic_8(Dcyclic_8(0:7))
}
