% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih12.R
\name{is_Dih12}
\alias{is_Dih12}
\title{Check whether an sgroupr_sgrp is Dih12}
\usage{
is_Dih12(x)
}
\arguments{
\item{x}{An sgroupr_sgrp}
}
\value{
True or False
}
\description{
Check whether an sgroupr_sgrp is Dih12
}
\examples{
x <- Dih12(0:11)
is_Dih12(x)
}
