% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih8_square.R
\name{is_Dih8}
\alias{is_Dih8}
\title{Test whether x is Dih8}
\usage{
is_Dih8(x)
}
\arguments{
\item{x}{A Object of Sgroupr_sgrp}
}
\value{
True or False
}
\description{
Test whether x is Dih8
}
\examples{
x <- Dih8(0:7)
is_Dih8(x)
}
