% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_12.R
\name{is_Dih_12}
\alias{is_Dih_12}
\title{Check whether an sgroupr_sgrp is Dih_12}
\usage{
is_Dih_12(x)
}
\arguments{
\item{x}{An sgroupr_sgrp}
}
\value{
True or False
}
\description{
Check whether an sgroupr_sgrp is Dih_12
}
\examples{
x <- Dih_12(0:11)
is_Dih_12(x)
}
