% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_6.R
\name{is_Dih_6}
\alias{is_Dih_6}
\title{Test whether x is Dih_6}
\usage{
is_Dih_6(x)
}
\arguments{
\item{x}{A Object of Sgroupr_sgrp}
}
\value{
True or False
}
\description{
Test whether x is Dih_6
}
\examples{
x <- Dih_6(0:5)
is_Dih_6(x)
}
