% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_8.R
\name{is_Dih_8}
\alias{is_Dih_8}
\title{Check if a sgroupr_sgpr is dihedral group Dih8 (Heisenberg)}
\usage{
is_Dih_8(x)
}
\arguments{
\item{x}{An sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is dihedral group Dih8 (Heisenberg)
}
\examples{
is_Dih_8(Dih_8(0:7))
}
