% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_8_square.R
\name{is_Dih_8_square}
\alias{is_Dih_8_square}
\title{Test whether x is Dih_8 (Square)}
\usage{
is_Dih_8_square(x)
}
\arguments{
\item{x}{A Object of Sgroupr_sgrp}
}
\value{
True or False
}
\description{
Test whether x is Dih_8 (Square)
}
\examples{
x <- Dih_8_square(0:7)
is_Dih_8_square(x)
}
