% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K4_C3.R
\name{is_K4_C3}
\alias{is_K4_C3}
\title{Test whether x is K4_C3}
\usage{
is_K4_C3(x)
}
\arguments{
\item{x}{A Object of Sgroupr_sgrp}
}
\value{
True or False
}
\description{
Test whether x is K4_C3
}
\examples{
x <- K4_C3(0:11)
is_K4_C3(x)
}
