% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_4.R
\name{is_Klein_4}
\alias{is_Klein_4}
\title{Check if a sgroupr_sgpr is Klein_4}
\usage{
is_Klein_4(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Klein_4
}
\examples{
is_Klein_4(Klein_4(0:3))
}
