% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_8 group.R
\name{is_Klein_8}
\alias{is_Klein_8}
\title{Check if a sgroupr_sgpr is Klein_8}
\usage{
is_Klein_8(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Klein_8
}
\examples{
is_Klein_8(Klein_8(0:7))
}
