% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q8_C2.R
\name{is_Q8_C2}
\alias{is_Q8_C2}
\title{Check if a sgroupr_sgpr is Q8_C2}
\usage{
is_Q8_C2(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is Q8_C2
}
\examples{
is_Q8_C2(Q8_C2(0:15))
}
