% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{is_S3}
\alias{is_S3}
\title{Test whether x is S3}
\usage{
is_S3(x)
}
\arguments{
\item{x}{A Object of Sgroupr_sgrp}
}
\value{
True or False
}
\description{
Test whether x is S3
}
\examples{
x <- S3(0:7)
is_S3(x)
}
