% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrp.R
\name{is_sgrp}
\alias{is_sgrp}
\title{sgrp group Check}
\usage{
is_sgrp(x)
}
\arguments{
\item{x}{is an sgroupr_sgrp object}
}
\value{
A boolean value show whether or not the object is sgroupr_sgrp group.
}
\description{
sgrp group Check
}
\examples{
x <- sgrp(c(1,2,3,4))
is_sgrp(x)
}
