% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_4.R
\name{Klein_4}
\alias{Klein_4}
\title{Create a Klein_4 sgroupr_sgrp object}
\usage{
Klein_4(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group Klein_4
}
\description{
Klein_4 belongs to the Klein Group
}
\examples{
Klein_4(0:3)
}
