% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_8.R
\name{Klein_8}
\alias{Klein_8}
\title{Create a Klein_8 sgroupr_sgrp object}
\usage{
Klein_8(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group Klein_8
}
\description{
Create a Klein_8 sgroupr_sgrp object
}
\examples{
Klein_8(0:7)
}
