% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_4 group.R
\name{new_Cyclic_4}
\alias{new_Cyclic_4}
\title{Create Cyclic_4 sgroupr_sgrp object}
\usage{
new_Cyclic_4(x = integer())
}
\arguments{
\item{x, }{an integer vector}
}
\description{
the length of this group should be 3
}
