% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dicyclic group 8.R
\name{new_Dcyclic_8}
\alias{new_Dcyclic_8}
\title{Create Dicyclic 8 sgroupr_sgrp object}
\usage{
new_Dcyclic_8(x = integer())
}
\arguments{
\item{x, }{an integer vector}
}
\description{
the length of this group should be 8
}
