% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klein_8 group.R
\name{new_Klein_8}
\alias{new_Klein_8}
\title{Create Klein_8 sgroupr_sgrp object}
\usage{
new_Klein_8(x = integer())
}
\arguments{
\item{x, }{an integer vector}
}
\description{
the length of this group should be 8
}
