% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q8_C2.R
\name{Q8C2}
\alias{Q8C2}
\title{Create a Q8_C2 sgroupr_sgrp object}
\usage{
Q8C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group Q8_C2
}
\description{
Create a Q8_C2 sgroupr_sgrp object
}
\examples{
Q8C2(0:15)
}
