% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{S3}
\alias{S3}
\title{Create a S3 small group of sgroupr_sgrp}
\usage{
S3(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
Return an object of sgroupr_sgrp with group S3
}
\description{
Create a S3 small group of sgroupr_sgrp
}
\examples{
S3(0:7)
}
