% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtcrs package.R
\docType{package}
\name{sgroupr-vctrs}
\alias{sgroupr}
\alias{sgroupr-package}
\title{Internal vctrs methods}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
