% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table}
\alias{table}
\title{Cayley Table}
\usage{
table(x)
}
\arguments{
\item{x}{An sgroupr_sgrp object}
}
\value{
A matrix for arithmetic operation
}
\description{
Cayley Table
}
\examples{
x <- Cyclic_4(0:3)
table(x)
}
