% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{vec_cast.sgroupr_sgrp.sgroupr_sgrp}
\alias{vec_cast.sgroupr_sgrp.sgroupr_sgrp}
\title{Casting}
\usage{
\method{vec_cast}{sgroupr_sgrp.sgroupr_sgrp}(x, to, ...)
}
\arguments{
\item{x}{An object}

\item{to}{An object}

\item{...}{Other parameters}
}
\value{
A sgroupr object
}
\description{
Allows explicitly to change the object type
}
