% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{vec_ptype_abbr.sgroupr_sgrp}
\alias{vec_ptype_abbr.sgroupr_sgrp}
\title{Coercion}
\usage{
\method{vec_ptype_abbr}{sgroupr_sgrp}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Other parameters}
}
\value{
A character of abbreviation
}
\description{
Allows implicitly to change the object type
}
