% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_12.R
\name{as_Dih_12}
\alias{as_Dih_12}
\title{Change a vector into object of Dih8 group}
\usage{
as_Dih_12(x)
}
\arguments{
\item{x}{A vector}
}
\value{
An Sgroupr_sgrp group belonging to Dih8
}
\description{
Change a vector into object of Dih8 group
}
\examples{
x <- 0:7
as_Dih_12(x)
}
