% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrp.R
\name{as_sgrp}
\alias{as_sgrp}
\title{Chang a vector into sgroupr_sgrp object}
\usage{
as_sgrp(x, ...)
}
\arguments{
\item{x}{A sgroupr_sgrp object}

\item{...}{other params}
}
\description{
Chang a vector into sgroupr_sgrp object
}
\examples{
as_sgrp(c(1L,2L,3L,4L),group = "Dih_6")
}
