% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C2.R
\name{C4_C2}
\alias{C4_C2}
\title{Create a C4C2 sgroupr_sgrp object}
\usage{
C4_C2(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a sgroupr_sgpr object with group C4C2
}
\description{
Create a C4C2 sgroupr_sgrp object
}
\examples{
C4_C2(0:7)
}
