% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_C3.R
\name{C4_C3}
\alias{C4_C3}
\title{Create a Group C4_C3 sgroupr_sgrp object}
\usage{
C4_C3(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with group C4_C3
}
\description{
C4_C3 belongs to the multiplication Group
}
\examples{
C4_C3(0:11)
}
