% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclic_16.R
\name{Cyclic_16}
\alias{Cyclic_16}
\title{Create a Cyclic 16 sgroupr_sgrp object}
\usage{
Cyclic_16(x = integer())
}
\arguments{
\item{x}{Is an integer vector}
}
\value{
A sgroupr_sgpr object with group Cyclic 16
}
\description{
The length of Cyclic 16 should be 0 to 15
}
\examples{
Cyclic_16(0:15)
}
