% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_12.R
\name{Dih_12}
\alias{Dih_12}
\title{Helper function of Dih2 group}
\usage{
Dih_12(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
a Dih_12 sgroupr_sgrp
}
\description{
Helper function of Dih2 group
}
\examples{
Dih_12(0:11)
}
