% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dih_8_square.R
\name{Dih_8_square}
\alias{Dih_8_square}
\title{Create a Dih_8 (Square) small group of sgroupr_sgrp}
\usage{
Dih_8_square(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
An object of sgroupr_sgrp with group Dih_8
}
\description{
Dih_8 (Square) belongs to the dihedral Group
}
\examples{
Dih_8_square(0:7)
}
