% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Extract the group attribute of a sgroupr_sgrp}
\usage{
group(x)
}
\arguments{
\item{x}{An sgroupr_sgrp object}
}
\value{
a character in lenth of 1.
}
\description{
Extract the group attribute of a sgroupr_sgrp
}
\examples{
x <- Klein_4(c(0,1,2,3,4,5,6,7))
group(x)
}
