% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3C2 group.R
\name{is_C3C2}
\alias{is_C3C2}
\title{Check if a sgroupr_sgpr is C3C2}
\usage{
is_C3C2(x)
}
\arguments{
\item{x}{an sgroupr_sgpr object}
}
\value{
True of False
}
\description{
Check if a sgroupr_sgpr is C3C2
}
\examples{
is_C3C2(C6C2(0:5))
}
