% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K4_C3.R
\name{K4_C3}
\alias{K4_C3}
\title{Create a K4_C3 small group of sgroupr_sgrp}
\usage{
K4_C3(x = integer())
}
\arguments{
\item{x}{An integer vector}
}
\value{
An object of sgroupr_sgrp with group K4_C3
}
\description{
K4_C3 belongs to the Multiplication Group
}
\examples{
K4_C3(0:11)
}
