% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpr.R
\name{sgrp}
\alias{sgrp}
\alias{is_sgrp}
\title{Genealized small group}
\usage{
sgrp(x = integer(), group = NA_character_)

is_sgrp(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{is_sgrp()}: An object to test.
\item For \code{sgrp()}: A vector
}}

\item{group}{}
}
\value{
An S3 vector of class \code{sgroupr_sgrp}.
}
\description{
Genealized small group
}
\examples{
sgrp(c(1,2,3,4))

}
