% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrp.R
\name{sgrp}
\alias{sgrp}
\title{Generalized small group}
\usage{
sgrp(x = integer(), group = NA_character_)
}
\arguments{
\item{x}{\itemize{
\item For \code{is_sgrp()}: An object to test.
\item For \code{sgrp()}: A vector
}}

\item{group}{a chacter which shows which small group x is.}
}
\value{
An S3 vector of class \code{sgroupr_sgrp}.
}
\description{
Generalized small group
}
\examples{
sgrp(c(1,2,3,4))
}
