% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Cayley table for finite small group}
\usage{
table(x)
}
\arguments{
\item{x}{An sgroupr_sgrp object}
}
\value{
A matrix for arithmetic operation
}
\description{
Cayley table for finite small group
}
\examples{
x <- S3(0:5)
table(x)
}
