% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arithmetic operations.R
\name{vec_arith.sgroupr_sgrp}
\alias{vec_arith.sgroupr_sgrp}
\title{Arithmetic Operations}
\usage{
vec_arith.sgroupr_sgrp(op, x, y, ...)
}
\arguments{
\item{op}{x
y}
}
\description{
This script contains the arithmetic operations
}
