% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{vec_arith.sgroupr_sgrp}
\alias{vec_arith.sgroupr_sgrp}
\title{Title}
\usage{
vec_arith.sgroupr_sgrp(op, x, y, ...)
}
\arguments{
\item{op}{operation such as \code{+}}

\item{x}{an sgrp}

\item{y}{another sgrp}

\item{...}{Other params}
}
\value{
a result
}
\description{
Title
}
