% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{vec_ptype_abbr.sgroupr_sgrp}
\alias{vec_ptype_abbr.sgroupr_sgrp}
\title{Coercion file created for small group}
\usage{
\method{vec_ptype_abbr}{sgroupr_sgrp}(x, ...)
}
\description{
Coercion file created for small group
}
